AddMovie now working
AddMovie is now working.
Had two major issues before managing to make it work:
If I type the values which I send to the service I get an error. String not valid for String|Null . So I had to revert to using <Any> type.
Second issue is that I stringified the body, seems like I didnt need to do that >:(
Also angular crashes, if I dont explicity tell the response is text, if the reponse isnt Json format.